Sella Ai/requirements.txt
telethon
ctransformers
aiosqlite
googlesearch-python
python-dotenv 